requirements.txt
# 核心框架
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.6

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 数据处理
pandas>=2.0.0
openpyxl>=3.1.0
tabulate>=0.9.0

# 图像处理
Pillow>=10.2.0
opencv-python>=4.9.0

# PDF处理
PyMuPDF>=1.23.0

# 语义分片
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# 日志和监控
loguru>=0.7.0

# 测试
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.26.0

# MinerU (核心依赖)
# 注意：MinerU需要单独安装，因为它有很多依赖
# pip install "magic-pdf[full]>=0.6.0"
